+++ rust/src/libponyc/pass/completeness.rs
@@ -821,7 +821,7 @@ unsafe extern "C" fn completeness_match(mut _opt: *mut pass_opt_t, mut ast: *mut
             .wrapping_sub(1 as libc::c_int as libc::c_ulong),
         children.as_mut_ptr(),
     );
-    let mut branch_count: size_t = 0 as libc::c_int as size_t;
+    let mut branch_count: size_t = 0;
     if ast_checkflag(cases, AST_FLAG_JUMPS_AWAY as libc::c_int as uint32_t) == 0 {
         branch_count = branch_count.wrapping_add(1);
         ast_inheritbranch(ast, cases);
